+++ private/test/src/matrix-cerbos/admin.test.ts
@@ -232,6 +232,7 @@ describe("Client", () => {
 
           const outputs: OutputResult[] = [
             {
+              action: versionIsAtLeast("0.51.0", cerbosVersion) ? "delete" : "",
               source: "resource.document.v1#delete",
               value: "delete_allowed:me@example.com",
             },